Rename test script file
